+++ packages/stacks-svelte/src/components/Avatar/Avatar.svelte
@@ -49,6 +49,11 @@
          * Additional CSS classes added to the element
          */
         class?: ClassValue;
+
+        /**
+         * Localized translation for private communities icon title tag
+         */
+        i18nPrivateIconTitle?: string;
     }
 
     const {
@@ -61,6 +66,7 @@
         badge = false,
         class: className = "",
         role,
+        i18nPrivateIconTitle = "Private",
         ...restProps
     }: Props & HTMLAnchorAttributes = $props();
 
@@ -81,7 +87,14 @@
     {...restProps}
 >
     {#if src}
-        <img class="s-avatar--image" {src} alt="" role="presentation" />
+        <img
+            class="s-avatar--image"
+            {src}
+            alt=""
+            role="presentation"
+            width={size}
+            height={size}
+        />
     {:else if letter}
         <span class="s-avatar--letter" aria-hidden="true">{letter}</span>
     {/if}
@@ -95,8 +108,11 @@
         />
     {/if}
     {#if badge}
-        <!-- TODO This badge is not purely decorative, so it should include descriptive text
-        (see https://stackoverflow.atlassian.net/browse/A11Y-126) -->
-        <Icon class="s-avatar--badge" src={IconShieldXSm} native />
+        <Icon
+            class="s-avatar--badge"
+            src={IconShieldXSm}
+            native
+            title={i18nPrivateIconTitle}
+        />
     {/if}
 </svelte:element>